add bags in the work

+++ src/main/java/ru/javawebinar/topjava/service/MealToServiceImpl.java
@@ -8,8 +8,8 @@ import java.util.List;
 public interface MealToServiceImpl {
     List<Meal> loadMeals();
     //void DeletedMeals();
-    Meal getMealsById(long id);
-    void deleteMealsById(long id);
+    Meal getMealsById(int id);
+    void deleteMealsById(int id);
     Meal updateUserData(int id, LocalDateTime dateTime, String description, int calories);
-     int returnCaloriesNorm();
+    int returnCaloriesNorm();
 }
